+++ ForestBlog/src/main/java/com/liuyanzhao/blog/entity/ArticleTagRef.java
@@ -14,14 +14,14 @@ import java.io.Serializable;
 public class ArticleTagRef implements Serializable {
     private static final long serialVersionUID = -5816783232020910492L;
 
-    private Integer articleId;
+    private Long articleId;
 
     private Integer tagId;
 
     public ArticleTagRef() {
     }
 
-    public ArticleTagRef(Integer articleId, Integer tagId) {
+    public ArticleTagRef(Long articleId, Integer tagId) {
         this.articleId = articleId;
         this.tagId = tagId;
     }
